Visual-Scripts Breakdown
* Move the webpack loader into new visual-entry-loader package.
* Move the oss-report-builder into a new package.
* Add a package that defines the common configuration-loading logic between the entry-loader and the scripts.
