--- packages/create-powerbi-visual/empty-visual/pbiviz.json
@@ -1,24 +1,0 @@
-{
-  "visual": {
-    "name": "<%= name %>",
-    "displayName": "<%= name %>",
-    "guid": "<%= name %><%= guid %>",
-    "visualClassName": "Visual",
-    "description": "",
-    "supportUrl": "",
-    "gitHubUrl": "",
-    "entry": "src/visual.ts"
-  },
-  "apiVersion": "1.7.0",
-  "author": {
-    "name": "",
-    "email": ""
-  },
-  "assets": {
-    "icon": "assets/icon.png"
-  },
-  "style": "style/visual.less",
-  "capabilities": "capabilities.json",
-  "dependencies": "dependencies.json",
-  "stringResources": []
-}
