+++ packages/visual-entry-loader/package.json
@@ -1,0 +1,11 @@
+{
+	"name": "@essex/visual-entry-loader",
+	"description":
+		"A webpack loader for injecting powerbi visual bootstrapping logic around a custom visual module",
+	"version": "0.0.1",
+	"main": "src/index.js",
+	"license": "MIT",
+	"dependencies": {
+		"@essex/visual-config": "^0.0.1"
+	}
+}
