+++ packages/visual-scripts/package.json
@@ -1,39 +1,42 @@
 {
-    "name": "@essex/powerbi-visual-scripts",
-    "version": "2.0.2",
-    "publishConfig": {
-        "access": "public"
-    },
-    "bin": {
-        "essex-powerbi-visual-scripts": "src/essex-pbi-scripts.js"
-    },
-    "license": "MIT",
-    "author": "Microsoft",
-    "dependencies": {
-        "@types/chai": "^4.0.4",
-        "@types/d3": "^4.10.0",
-        "@types/mocha": "^2.2.42",
-        "chokidar": "^1.7.0",
-        "clean-css": "^4.1.7",
-        "commander": "^2.11.0",
-        "connect": "^3.6.3",
-        "css-loader": "^0.28.5",
-        "d3": "^4.10.0",
-        "exports-loader": "^0.6.4",
-        "expose-loader": "^0.7.3",
-        "ignore-loader": "^0.1.2",
-        "json-loader": "^0.5.7",
-        "lodash": "^4.17.4",
-        "memory-fs": "^0.4.1",
-        "mkdirp": "^0.5.1",
-        "node-sass": "^4.5.3",
-        "node-zip": "^1.1.1",
-        "raw-loader": "^0.5.1",
-        "sass-loader": "^6.0.6",
-        "script-loader": "^0.7.0",
-        "serve-static": "^1.12.4",
-        "style-loader": "^0.19.0",
-        "ts-loader": "^2.3.7",
-        "webpack": "^3.5.5"
-    }
+	"name": "@essex/powerbi-visual-scripts",
+	"version": "2.0.2",
+	"publishConfig": {
+		"access": "public"
+	},
+	"bin": {
+		"essex-powerbi-visual-scripts": "src/essex-pbi-scripts.js"
+	},
+	"license": "MIT",
+	"author": "Microsoft",
+	"dependencies": {
+		"@essex/oss-report-builder": "^0.0.1",
+		"@essex/visual-entry-loader": "^0.0.1",
+		"@essex/visual-config": "^0.0.1",
+		"@types/chai": "^4.0.4",
+		"@types/d3": "^4.10.0",
+		"@types/mocha": "^2.2.42",
+		"chokidar": "^1.7.0",
+		"clean-css": "^4.1.7",
+		"commander": "^2.11.0",
+		"connect": "^3.6.3",
+		"css-loader": "^0.28.5",
+		"d3": "^4.10.0",
+		"exports-loader": "^0.6.4",
+		"expose-loader": "^0.7.3",
+		"ignore-loader": "^0.1.2",
+		"json-loader": "^0.5.7",
+		"lodash": "^4.17.4",
+		"memory-fs": "^0.4.1",
+		"mkdirp": "^0.5.1",
+		"node-sass": "^4.5.3",
+		"node-zip": "^1.1.1",
+		"raw-loader": "^0.5.1",
+		"sass-loader": "^6.0.6",
+		"script-loader": "^0.7.0",
+		"serve-static": "^1.12.4",
+		"style-loader": "^0.19.0",
+		"ts-loader": "^2.3.7",
+		"webpack": "^3.5.5"
+	}
 }
